refactor: 재사용되는 ReviewShape 타입 파일분리
