+++ src/types/apiTypes.ts
@@ -1,0 +1,12 @@
+export type ReviewShape = {
+  id: string;
+  author: {
+    username: string;
+    profileImage: string;
+  };
+  content: string;
+  rating: number;
+  menu: string;
+  updatable: boolean;
+};
+
